Avoid depending on Keys.CONTROL as it means something different on MacOS
Running the test suite on MacOS before this change results in failures
like the following:

test_common.py::TestFuncSeleniumCommonChrome::test_fill - AssertionError: assert 'RockNew name' == 'New name'

This is because the 'name' input is not emptied before we attempt to
fill it. That is because the key "CONTROL" does not do the same thing
on MacOS as it does on other platforms.

Retaining the policy of not using elem.clear(), replace this with
a hopefully more robust method to select the contents of the field.